python/pyproject.toml
[tool.poetry]
name = "centrex-motorized-linear-stage"
version = "0.1.0"
description = ""
authors = ["ograsdijk <o.grasdijk@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
